requirements.txt
paho-mqtt>=2.1.0
customtkinter>=5.2.2
darkdetect>=0.8.0
packaging>=25.0